+++ packages/apollo-env/src/index.ts
@@ -1,4 +1,5 @@
 import "./polyfills";
-import "./typescript-utility-types";
 
+export * from "./typescript-utility-types";
 export * from "../lib/fetch";
+export * from "./utils";