+++ packages/apollo-env/src/typescript-utility-types.ts
@@ -1,2 +1,2 @@
-type WithRequired<T, K extends keyof T> = T & Required<Pick<T, K>>;
-type DeepPartial<T> = { [P in keyof T]?: DeepPartial<T[P]> };
+export type WithRequired<T, K extends keyof T> = T & Required<Pick<T, K>>;
+export type DeepPartial<T> = { [P in keyof T]?: DeepPartial<T[P]> };