+++ packages/apollo-graphql/src/index.ts
@@ -1,0 +1,5 @@
+export {
+  defaultEngineReportingSignature,
+  defaultOperationRegistrySignature,
+  operationHash
+} from "./operationId";